+++ Sources/Views/ALKFormCell.swift
@@ -167,6 +167,7 @@ class ALKFormCell: ALKChatBaseCell<ALKMessageViewModel>, UITextFieldDelegate, UI
         itemListView.register(ALKFormDateItemCell.self)
         itemListView.register(ALKFormTimeItemCell.self)
         itemListView.register(ALKFormDateTimeItemCell.self)
+        itemListView.register(KMFormDropDownCell.self)
     }
 
     private func setUpSubmitButton(title: String) {
@@ -382,6 +383,13 @@ extension ALKFormCell: UITableViewDataSource, UITableViewDelegate {
                 cell.valueTextField.text = ""
             }
             return cell
+        case .dropdown:
+            print("Drop Down Support ")
+            let cell: KMFormDropDownCell = tableView.dequeueReusableCell(forIndexPath: indexPath)
+            cell.item = item
+            cell.menu.tag = indexPath.section
+            cell.delegate = self
+            return cell
         }
     }
 
@@ -439,6 +447,19 @@ extension ALKFormCell: ALKDatePickerButtonClickProtocol {
         formSubmitData.dateFields[position] = timeInMillSecs
         formData = formSubmitData
         itemListView.reloadSections([position], with: .fade)
+    }
+}
+
+extension ALKFormCell: KMFormDropDownSelectionProtocol {
+    func optionSelected(position: Int, selectedText: String) {
+        guard let formSubmittedData = formData,
+              position < itemListView.numberOfSections
+        else {
+            print("Can't be updated due to incorrect index")
+            return
+        }
+        formSubmittedData.dropDownFields[position] = selectedText
+        formData = formSubmittedData
     }
 }
 
@@ -514,5 +535,6 @@ class FormDataSubmit {
     var singleSelectFields = [Int: Int]()
     var multiSelectFields = [Int: [Int]]()
     var dateFields = [Int: Int64]()
+    var dropDownFields = [Int: String]()
     var validationFields = [Int: Int]()
 }