+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 The changelog for [KommunicateChatUI-iOS-SDK](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK). Also see the [releases](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK/releases) on Github.
 
+## [Unreleased]
+- [CM-1451] Fixed issue in ALKTemplateMessage flow 
 ## [1.0.9] 2023-06-07
 - Added Customization to restrict browser navigation on tap of link list template
 - Fixed the upload issue for custom cloud support