+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # CHANGELOG
 
 The changelog for [KommunicateChatUI-iOS-SDK](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK). Also see the [releases](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK/releases) on Github.
+
 ## [Unreleased]
 - Added Cusotmization support for Faq button
+- [CM-1451] Fixed issue in ALKTemplateMessage flow 
+
 ## [1.0.9] 2023-06-07
 - Added Customization to restrict browser navigation on tap of link list template
 - Fixed the upload issue for custom cloud support